+++ CHANGELOG.md
@@ -1,12 +1,78 @@
 # Luma | Routing Component Changelog
 
+## [1.4.1] - 2024-04-17
+### Added
+- N/A
+
+### Changed
+- N/A
+
+### Deprecated
+- N/A
+
+### Removed
+- N/A
+
+### Fixed
+- Routes now properly considering defined methods
+
+### Security
+- N/A
+
 ## [1.4.0] - 2024-04-17
-- Allow controller methods to use the current `Request` as the first parameter
+### Added
+- N/A
+
+### Changed
+- Controller methods can now optionally use the current `Request` by typehinting the **first** parameter with `Request`
+
+### Deprecated
+- N/A
+
+### Removed
+- Removed Jenkins pipeline
+
+### Fixed
+- N/A
+
+### Security
+- N/A
 
 ## [1.3.9] - 2024-03-17
-- Clean up unit tests 
+### Added
+- N/A
+
+### Changed
+- N/A
+
+### Deprecated
+- N/A
+
+### Removed
+- N/A
+
+### Fixed
+- N/A
+
+### Security
+- Additional unit tests 
 
 ## [1.3.8] - 2024-03-02
-- Add build pipeline
-- Update unit tests
-- Add CHANGELOG
+### Added
+- Added build pipeline
+- Added CHANGELOG
+
+### Changed
+- N/A
+
+### Deprecated
+- N/A
+
+### Removed
+- N/A
+
+### Fixed
+- N/A
+
+### Security
+- Added additional unit tests